+++ poetry/console/commands/lock.py
@@ -2,11 +2,9 @@ from .env_command import EnvCommand
 
 
 class LockCommand(EnvCommand):
-    """
-    Locks the project dependencies.
 
-    lock
-    """
+    name = "lock"
+    description = "Locks the project dependencies."
 
     help = """The <info>lock</info> command reads the <comment>pyproject.toml</> file from
 the current directory, processes it, and locks the depdencies in the <comment>poetry.lock</> file.