+++ poetry/masonry/builders/__init__.py
@@ -1,3 +1,4 @@
 from .complete import CompleteBuilder
+from .editable import EditableBuilder
 from .sdist import SdistBuilder
 from .wheel import WheelBuilder