+++ poetry/masonry/builders/wheel.py
@@ -67,6 +67,10 @@ class WheelBuilder(Builder):
 
         (fd, temp_path) = tempfile.mkstemp(suffix=".whl")
 
+        st_mode = os.stat(temp_path).st_mode
+        new_mode = normalize_file_permissions(st_mode)
+        os.chmod(temp_path, new_mode)
+
         with zipfile.ZipFile(
             os.fdopen(fd, "w+b"), mode="w", compression=zipfile.ZIP_DEFLATED
         ) as zip_file:
@@ -80,7 +84,9 @@ class WheelBuilder(Builder):
             wheel_path.unlink()
         shutil.move(temp_path, str(wheel_path))
 
-        self._io.write_line(" - Built <fg=cyan>{}</>".format(self.wheel_filename))
+        self._io.write_line(
+            " - Built <comment>{}</comment>".format(self.wheel_filename)
+        )
 
     def _build(self, wheel):
         if self._package.build:
@@ -127,6 +133,9 @@ class WheelBuilder(Builder):
         to_add = []
 
         for include in self._module.includes:
+            if include.formats and "wheel" not in include.formats:
+                continue
+
             include.refresh()
 
             for file in include.elements: