+++ poetry/repositories/pypi_repository.py
@@ -57,7 +57,6 @@ class PyPiRepository(Repository):
     CACHE_VERSION = parse_constraint("0.12.0")
 
     def __init__(self, url="https://pypi.org/", disable_cache=False, fallback=True):
-        self._name = "PyPI"
         self._url = url
         self._disable_cache = disable_cache
         self._fallback = fallback
@@ -79,6 +78,16 @@ class PyPiRepository(Repository):
         )
 
         super(PyPiRepository, self).__init__()
+
+        self._name = "PyPI"
+
+    @property
+    def url(self):  # type: () -> str
+        return self._url
+
+    @property
+    def authenticated_url(self):  # type: () -> str
+        return self._url
 
     def find_packages(
         self,
@@ -105,7 +114,14 @@ class PyPiRepository(Repository):
             ):
                 allow_prereleases = True
 
-        info = self.get_package_info(name)
+        try:
+            info = self.get_package_info(name)
+        except PackageNotFound:
+            self._log(
+                "No packages found for {} {}".format(name, str(constraint)),
+                level="debug",
+            )
+            return []
 
         packages = []
 
@@ -419,6 +435,13 @@ class PyPiRepository(Repository):
 
             if info:
                 return info
+
+            # Prefer non platform specific wheels
+            if universal_python3_wheel:
+                return self._get_info_from_wheel(universal_python3_wheel)
+
+            if universal_python2_wheel:
+                return self._get_info_from_wheel(universal_python2_wheel)
 
             if platform_specific_wheels and "sdist" not in urls:
                 # Pick the first wheel available and hope for the best