+++ tests/masonry/builders/test_wheel.py
@@ -87,46 +87,6 @@ def test_wheel_module_src():
         assert "module_src.py" in z.namelist()
 
 
-def test_package_with_include(mocker):
-    # Patch git module to return specific excluded files
-    p = mocker.patch("poetry.vcs.git.Git.get_ignored_files")
-    p.return_value = [
-        str(
-            Path(__file__).parent
-            / "fixtures"
-            / "with-include"
-            / "extra_dir"
-            / "vcs_excluded.txt"
-        ),
-        str(
-            Path(__file__).parent
-            / "fixtures"
-            / "with-include"
-            / "extra_dir"
-            / "sub_pkg"
-            / "vcs_excluded.txt"
-        ),
-    ]
-    module_path = fixtures_dir / "with-include"
-    WheelBuilder.make(Poetry.create(str(module_path)), NullEnv(), NullIO())
-
-    whl = module_path / "dist" / "with_include-1.2.3-py3-none-any.whl"
-
-    assert whl.exists()
-
-    with zipfile.ZipFile(str(whl)) as z:
-        names = z.namelist()
-        assert len(names) == len(set(names))
-        assert "with_include-1.2.3.dist-info/LICENSE" in names
-        assert "extra_dir/__init__.py" in names
-        assert "extra_dir/vcs_excluded.txt" in names
-        assert "extra_dir/sub_pkg/__init__.py" in names
-        assert "extra_dir/sub_pkg/vcs_excluded.txt" not in names
-        assert "my_module.py" in names
-        assert "notes.txt" in names
-        assert "package_with_include/__init__.py" in names
-
-
 def test_dist_info_file_permissions():
     module_path = fixtures_dir / "complete"
     WheelBuilder.make(Poetry.create(str(module_path)), NullEnv(), NullIO())