+++ CHANGELOG.md
@@ -9,6 +9,9 @@
 - Added a `env use` command to control the Python version used by the project.
 - Added a `env list` command to list the virtualenvs associated with the current project.
 - Added a `env remove` command to delete virtualenvs associated with the current project.
+- Added support for declaring a specific source for dependencies.
+- Added support for disabling PyPI and making another repository the default one.
+- Added support for declaring private repositories as secondary.
 
 ### Changed
 
@@ -22,6 +25,41 @@
 ### Fixed
 
 - Fixed transitive extra dependencies being removed when updating a specific dependency.
+
+
+## [0.12.16] - 2019-05-17
+
+### Fixed
+
+- Fixed packages with no hashes retrieval for legacy repositories.
+- Fixed multiple constraints for dev dependencies.
+- Fixed dependency resolution failing on badly formed package versions instead of skipping.
+- Fixed permissions of built wheels.
+>>>>>>> master
+
+
+## [0.12.15] - 2019-05-03
+
+### Fixed
+
+- Fixed an `AttributeError` in the editable builder.
+- Fixed resolution of packages with only Python 3 wheels and sdist when resolving for legacy repositories.
+- Fixed non-sha256 hashes retrieval for legacy repositories.
+
+
+## [0.12.14] - 2019-04-26
+
+### Fixed
+
+- Fixed root package installation for pure Python packages.
+
+
+## [0.12.13] - 2019-04-26
+
+### Fixed
+
+- Fixed root package installation with `pip>=19.0`.
+- Fixed packages not being removed after using the `remove` command.
 
 
 ## [0.12.12] - 2019-04-11
@@ -661,7 +699,11 @@ Initial release
 
 
 
-[Unreleased]: https://github.com/sdispater/poetry/compare/0.12.12...develop
+[Unreleased]: https://github.com/sdispater/poetry/compare/0.12.16...develop
+[0.12.16]: https://github.com/sdispater/poetry/releases/tag/0.12.16
+[0.12.15]: https://github.com/sdispater/poetry/releases/tag/0.12.15
+[0.12.14]: https://github.com/sdispater/poetry/releases/tag/0.12.14
+[0.12.13]: https://github.com/sdispater/poetry/releases/tag/0.12.13
 [0.12.12]: https://github.com/sdispater/poetry/releases/tag/0.12.12
 [0.12.11]: https://github.com/sdispater/poetry/releases/tag/0.12.11
 [0.12.10]: https://github.com/sdispater/poetry/releases/tag/0.12.10